requirements.txt
click==8.1.7
dnspython==2.7.0
gensim==4.3.3
joblib==1.4.2
nltk==3.9.1
numpy==1.26.4
pymongo==4.10.1
python-dotenv==1.0.1
regex==2024.9.11
scipy==1.13.1
smart-open==7.0.5
tqdm==4.66.5
wrapt==1.16.0
